split out s3 functions in its own helper
and a spec
