+++ app/views/user_mailer/async_report_notification.html.erb
@@ -1,3 +1,4 @@
+<%= render 'user_mailer/components/header' %>
 <p>
     A Touchpoints export was initiated at <%= @start_time %>,
     and completed at
