+++ app/views/user_mailer/form_feedback.html.erb
@@ -1,13 +1,12 @@
+<%= render 'user_mailer/components/header' %>
 <p>
     You recently archived the form, <%= link_to @form.name, admin_form_url(@form) %>, in Touchpoints.
 </p>
-
 <p>
     The form was created <%= time_ago_in_words(@form.created_at) %> ago.
     <br>
     The form received <%= @form.response_count %> responses.
 </p>
-
 <p>
     Please take this <%= link_to "short survey", @feedback_url %>
     to provide feedback regarding the Touchpoints application and/or the Touchpoints team.
@@ -15,7 +14,6 @@
     <br>
     <%= link_to @feedback_url, @feedback_url %>
 </p>
-
 <p>
 We want to know what you thought of Touchpoints and how CX feedback is going.
 </p>
