+++ app/views/user_mailer/user_reactivation_email.html.erb
@@ -1,19 +1,15 @@
 <%= render 'user_mailer/components/header' %>
-
 <p>
   Hi <%= @user.email %>,
 </p>
-
 <p>
 Touchpoints accounts are automatically inactivated
 after 30 days without login, per security policy.
 But, your Touchpoints account has now been reactivated.
 </p>
-
 <p>
 Any questions, please us know,
 </p>
-
 <br>
 --
 <br>
